Integrate trick play.
- Add trick play adaptation in mpd. This addresses #178.

+++ packager/media/base/video_stream_info.cc
@@ -48,7 +48,7 @@ VideoStreamInfo::VideoStreamInfo(int track_id,
                                  uint16_t height,
                                  uint32_t pixel_width,
                                  uint32_t pixel_height,
-                                 int16_t trick_play_rate,
+                                 uint32_t trick_play_rate,
                                  uint8_t nalu_length_size,
                                  const std::string& language,
                                  bool is_encrypted)
