+++ packager/media/formats/wvm/wvm_media_parser.cc
@@ -574,7 +574,7 @@ bool WvmMediaParser::ParseIndexEntry() {
     }
 
     uint64_t track_duration = 0;
-    int16_t trick_play_rate = 0;
+    uint32_t trick_play_rate = 0;
     uint32_t sampling_frequency = kDefaultSamplingFrequency;
     uint32_t time_scale = kMpeg2ClockRate;
     uint16_t video_width = 0;
